Remove configuration->agent dependency
