+++ framework/wazuh/configuration.py
@@ -7,7 +7,6 @@ from xml.etree.ElementTree import fromstring
 from os import listdir, path as os_path
 import re
 from wazuh.exception import WazuhException
-from wazuh.agent import Agent
 from wazuh import common
 from wazuh.utils import cut_array
 
@@ -412,19 +411,12 @@ def get_ossec_conf(section=None, field=None):
     return data
 
 
-def get_agent_conf(group_id=None, offset=0, limit=common.database_limit, filename=None):
+def get_agent_conf_from_path(agent_conf, offset=0, limit=common.database_limit, filename=None):
     """
     Returns agent.conf as dictionary.
 
     :return: agent.conf as dictionary.
     """
-
-    if group_id:
-        if not Agent.group_exists(group_id):
-            raise WazuhException(1710, group_id)
-
-        agent_conf = "{0}/{1}".format(common.shared_path, group_id)
-
     if filename:
         agent_conf_name = filename
     else:
@@ -448,8 +440,8 @@ def get_agent_conf(group_id=None, offset=0, limit=common.database_limit, filenam
 
         # Parse XML to JSON
         data = _agentconf2json(xml_data)
-    except:
-        raise WazuhException(1101)
+    except Exception as e:
+        raise WazuhException(1101, str(e))
 
 
     return {'totalItems': len(data), 'items': cut_array(data, offset, limit)}
